+++ DuDoong-Api/build.gradle
@@ -16,8 +16,6 @@ dependencies {
 
     implementation project(':DuDoong-Domain')
     implementation project(':DuDoong-Common')
-    implementation project(':DuDoong-Infrastructure')
-
 }
 
 test {
